requirements.txt
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
avro-python3==1.10.2
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.9.7
fasteners==0.19
find_libpython==0.4.0
grpcio==1.68.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
kafka-python==2.0.2
mypy-extensions==1.0.0
numpy==1.24.4
objsize==0.6.1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pemja==0.4.1
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.16.0
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zstandard==0.23.0
